Added if statements to the moveRobot() method to determine the difference between the start and end coordinates. If there is a difference, the getTurn() method is launched, which turns the robot in the direction in which it will step.
Inside the getTurn() method there is a check if the final path is the nearest right neighbor, we call the turnRight() method, and for everything else turnLeft().
This is done in order to avoid unnecessary turns.

After getTurn() has been completed, the getMove() method is then run, into which we pass our object, the starting and ending points of the route.
The while loop will run the stepForward() method until the int variables passed inside the method are equal.
In order for the robot to move to both sides of the coordinate axis, an if check has been added, which makes an increment or decrement depending on whether the current position is greater or less than the end